+++ app/src/main/res/drawable-mdpi/ic_action_search.png
