+++ app/src/main/res/drawable-xhdpi/ic_action_search.png
