+++ app/src/main/res/values/styles.xml
@@ -1,0 +1,3 @@
+<resources>
+    <style name="AppTheme" parent="Theme.AppCompat.Light" />
+</resources>
