+++ gen/joonjeong/gammagraphhistogram/BuildConfig.java
@@ -1,6 +1,8 @@
-/** Automatically generated file. DO NOT MODIFY */
+/*___Generated_by_IDEA___*/
+
 package joonjeong.gammagraphhistogram;
 
+/* This stub is only used by the IDE. It is NOT the BuildConfig class actually packed into the APK */
 public final class BuildConfig {
-    public final static boolean DEBUG = true;
+  public final static boolean DEBUG = Boolean.parseBoolean(null);
 }
