--- res/values-v11/styles.xml
@@ -1,5 +1,0 @@
-<resources>
-
-    <style name="AppTheme" parent="android:Theme.Holo.Light" />
-
-</resources>
