--- res/values/styles.xml
@@ -1,5 +1,0 @@
-<resources>
-
-    <style name="AppTheme" parent="android:Theme.Light" />
-
-</resources>
